requirements.txt
dropbox>=11.36.0
assemblyai>=0.17.0
openai>=1.0.0
anthropic>=0.7.0
pyyaml>=6.0
requests>=2.28.0
ffmpeg-python>=0.2.0